+++ CHANGES.rst
@@ -3,6 +3,7 @@ Changes in 0.10.1 (2019-XX-XX)
 
 Bug fix:
  * Room cell: The states of direct chat and favorite buttons are reversed in the menu (#2788).
+ * Pasteboard: Fix a crash when passing a nil object to UIPasteboard.
 
 Changes in 0.10.0 (2019-10-11)
 ===============================================